recaped tq3940's work, by renaming files, and fixing folder structure
