+++ src/App.jsx
@@ -1,10 +1,9 @@
 import "./App.css";
-import Home from "./pages/Home/Home";
+import Home from "./pages/Home/";
 
 function App() {
   return (
     <>
-   
     <Home/>
     </>
   );
