--- src/components/counter/counter.jsx
@@ -1,15 +1,0 @@
-import { useState } from "react";
-const Counter = () => {
-    const [count, setCount] = useState(0)
-    return (
-        <div className="flex h-screen">
-            <div className="m-auto">
-                <div className="text-6xl text-red-800">{count}</div>
-                <button className="px-6 py-2 rounded bg-blue-800 hover:bg-pink-700 text-black" type="button" onClick={() => setCount((count) => count + 1)}>
-                    count+
-                </button>
-            </div>
-        </div>
-    )
-}
-export default Counter;
