+++ src/pages/Home/index.css
@@ -1,8 +1,10 @@
 @import url("https://fonts.googleapis.com/css2?family=Orbitron:wght@500&display=swap");
+
 :root {
   --yellow: #f6fe84;
   --white: #ffffff;
 }
+
 @media only screen and (min-width: 992px){
 
     .carousel-text-one {
