+++ tailwind.config.js
@@ -6,6 +6,9 @@ export default {
   ],
   theme: {
     extend: {},
+    colors: {
+      mainBgColor: '#191919'
+    },
   },
   plugins: [],
 }
